% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenInstituteBrainData-package.R
\docType{package}
\name{AllenInstituteBrainData-package}
\alias{AllenInstituteBrainData-package}
\alias{AllenInstituteBrainData}
\title{AllenInstituteBrainData}
\description{

}
\author{
Dario Righelli
}
