% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenInstituteBrainData.R
\name{AllenInstituteBrainData}
\alias{AllenInstituteBrainData}
\title{AllenInstituteBrainData}
\usage{
AllenInstituteBrainData(
  dataset = c("Allen_Mouse_2020"),
  verbose = TRUE,
  DataClass = c("SingleCellExperiment", "SummarizedExperiment")
)
}
\arguments{
\item{dataset}{the identifier of the Allen Brain dataset to retrieve}
}
\value{

}
\description{
AllenInstituteBrainData
}
\examples{
se <- AllenInstituteBrainData("Allen_Human_2020")
se

}
